CoreData_implement
Use core data to store the todo details and and i have set the attributes availble. and on popup save button press tittle and description added in the Todo List and on Swipe the Todo Item is deleted from the todo list. and when i have stop application and re_Run it and the item in Todo list i have added before already exist.This mean our local storage is working properly to manage our Todo List